TDD initial creation of R,P,S
Using TDD, created GameObjs - Rock, Paper and Scissors.
Beats method was added and overridden to address each combination for
each object. This is only a holding method - to be changed to address
any expansion of Lizard-Spock.

GameObj abstract class created and factory method implement using
reflection to obtain object. This mirrors the SML program written 2015.

Unknown object needs tidying up.

Tests also carried out for all test.

HumanObj created to choose the gameObj. This will eventually implement
Player. A new Class Computer will also implement Player, with additional
matter of AI controls to be addressed - this has as yet not been mapped.
It is expected that a factory pattern will also be used to obtain the
player type and in the case of the Computer - an AI level.