Added FatcoryPattern to obtain PlayerType
Due to only two types of player and cannot foresee this being extended -
then an a simplistic factory was used to obtain the playerTypes - via
PlayerInterface.

issue to content with is whether to apply the strategy pattern to the
Computer player with Intelligence at the point of creation or utilise an
Abstract pattern to obtain the intelligence level and inject it into
computer player later.