Added Computerintelligence Factory and GameObjectChoice Enum
Added game intelligence factory so that a level of intelligence can be
used by the computer.

Added Enum GameObjectChoice - so this can be extended later to include
spock and lizard. Also added static method of randomly choosing one of
the constants. This will be utilised by the novice computer.

Interface created for intelligence

TDD utilised to create Enum class and intelligence classes. Tests
carried out and complete