removed enum from intelligence and created playerfactory
made enum in a class of its own so that this will hopefully be the only
class that needs amending in order to add other objects
Playerfactory to obtain a type of player - human or computer - possibly
a factory is not required and the strategy pattern could be used without
need for casting from the factory pattern.

TDD carried out to create factory.

Also amended Intelligence and NoiceAI to conduct the random choice made
by the computer player.