Amended Player class to be abstract and added intelligence object
Changed the player interface to abstract class with the added extra of
implementing the strategy pattern using Intelligence intelLevel. This
way the computer can make a choice and how it makes a choice is
encapsulated.

However issue arises with human players - either throw exception or can
use this to prompt player to choose gameobject - to be determind