Added StateManager and two states to control the state of the game
Used StatePattern to control the movement of the state of the game.
StatePattern has GameState object that has an update function - this then allows the currentstate (GameState obj) to call its own updateState Method.

MenuState moves onto PlayState. and any other state that may need to move to... to be constructed as this progresses
Playstate update has not been tested and consequently created as yet

Intelligence has been amended - this needs a rethink